backend/requirements.txt
amqp==5.3.1
asgiref==3.10.0
async-timeout==5.0.1
attrs==25.4.0
billiard==4.2.2
black==25.9.0
celery==5.5.3
cfgv==3.4.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.10.7
cron_descriptor==2.0.6
distlib==0.4.0
Django==5.2.7
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-filter==25.2
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
filelock==3.20.0
identify==2.6.15
inflection==0.5.1
isort==7.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.0
pre_commit==4.3.0
prompt_toolkit==3.0.52
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.1.10
PyYAML==6.0.3
redis==6.4.0
referencing==0.37.0
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
tomli==2.3.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
virtualenv==20.35.3
wcwidth==0.2.14
